+++ extension/src/infrastructure/clients/chrome-guid-client.ts
@@ -1,6 +1,6 @@
 import type { GuidRepository } from "../../domain/guid/guid-repository";
-import { createMergeStore } from "./merge-store";
-import type { StorageArea } from "./storage-area";
+import { createMergeStore } from "../internal/merge-store";
+import type { StorageArea } from "../internal/storage-area";
 
 export function createChromeGuidClient(area: StorageArea): GuidRepository {
   return createMergeStore(area, "guids");
