+++ extension/src/infrastructure/internal/fetch-queue.ts
