+++ extension/src/infrastructure/internal/fixtures.ts
