+++ extension/src/infrastructure/internal/merge-store.ts
