+++ extension/src/infrastructure/internal/storage-area.ts
